requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
websocket-client==1.8.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
openai==1.3.7
speechrecognition==3.10.0
pyttsx3==2.90
pydub==0.25.1
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
transformers>=4.35.0
torch>=2.0.0
soundfile>=0.12.1
twilio>=8.0.0
# MongoDB for conversation storage
pymongo>=4.6.0
motor>=3.3.0
# Authentication dependencies
bcrypt>=4.0.0
PyJWT>=2.8.0

# Voice Agent specific dependencies
webrtcvad-wheels
